src/requirements.txt
Django==4.2.7
gunicorn==21.2.0
boto3==1.34.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.0
drf-spectacular==0.27.0
django-storages==1.14.2